rename params/neucogar to params/neuron_groups
