+++ memristive_spinal_cord/layer1/rybak/connectome.py
@@ -1,6 +1,6 @@
 import neucogar.namespaces as NEST_NAMESPACE
 from neucogar.SynapseModel import SynapseModel
-from memristive_spinal_cord.layer1.rybak.params.neucogar import *
+from memristive_spinal_cord.layer1.rybak.params.neuron_groups import *
 from memristive_spinal_cord.layer1.rybak.neuron_groups import *
 
 # synapses
