+++ memristive_spinal_cord/layer1/rybak/neuron_groups.py
@@ -1,6 +1,6 @@
 from neucogar.api_kernel import CreateNetwork
 from memristive_spinal_cord.layer1.neuron_group_network import NeuronGroupNetwork
-from memristive_spinal_cord.layer1.rybak.params.neucogar import Layer1Groups
+from memristive_spinal_cord.layer1.rybak.params.neuron_groups import Layer1Groups
 
 layer1_network = NeuronGroupNetwork()
