+++ memristive_spinal_cord/layer1/rybak/params/neuron_groups.py
